requirements.txt
ctransformers==0.2.5
sentence-transformers==2.2.2
flask
pypdf
python-dotenv
langchain-pinecone==0.0.3
pinecone-client==3.2.1
langchain==0.1.13
langchain-community==0.0.29
langchain-core==0.1.34
langchain-text-splitters==0.0.1
langchainplus-sdk==0.0.20
# pinecone-client
# langchain==0.0.225
